Remove unused cast

+++ drivers/builtin/src/chacha20.c
@@ -240,7 +240,7 @@ int mbedtls_chacha20_check_counter_wrap(const mbedtls_chacha20_context *ctx,
         size -= available_keystream;
     }
 
-    needed_blocks = (uint64_t) (size / MBEDTLS_CHACHA20_BLOCK_SIZE_BYTES);
+    needed_blocks = (size / MBEDTLS_CHACHA20_BLOCK_SIZE_BYTES);
 
     if (size % MBEDTLS_CHACHA20_BLOCK_SIZE_BYTES != 0U) {
         needed_blocks++;
